backend/pyproject.toml
[project]
name = "crucible-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.121.0",
  "uvicorn>=0.34.0",
  "httpx>=0.27.0",
  "psycopg2-binary>=2.9.9",
  "sqlalchemy>=2.0.0",
  "upstash-ratelimit>=1.1.0",
  "upstash-redis>=1.0.0",
  "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["backend*", "services*"]
exclude = ["public*", "migrations*"]
